core-svc/requirements.txt
# Core Service Requirements - Phase 1-5
# FastAPI backend with GPU rendering, Celery queue, and production monitoring

# Torch CUDA 12.2 compatibility (must be first)
--extra-index-url https://download.pytorch.org/whl/cu122

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery[redis]==5.3.4
redis==5.0.1

# Core Processing (from existing project)
openai>=1.3.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
python-dotenv>=1.0.0

# Audio/Video processing dependencies  
yt-dlp>=2023.11.16
pysrt>=1.1.2

# Speech VAD & NLP (Essential for filler word detection)
py-webrtcvad-wheels==2.0.10
spacy==3.8.7
fasttext>=0.9.2

# Deep Learning Stack (CUDA 12.2 compatible)
torch==2.2.*
torchaudio==2.2.*
torchvision==0.17.*

# Computer Vision (CPU-only, but fast)
opencv-python==4.11.*

# Face Detection & Landmark Models
mediapipe==0.10.9
mediapipe-tasks==0.10.4

# Object Tracking & Scene Detection
norfair==2.2.*
pyscenedetect==0.6.*

# Audio Processing (GPU accelerated)
soundfile>=0.12.1

# ONNX Runtime for RT-DETR export
onnxruntime-gpu==1.17.*

# RT-DETR v2 model (Phase 2)
ultralytics>=8.0.0

# SSIM calculation for visual quality metrics (Phase 2)
scikit-image>=0.21.0

# Utilities
pydantic==2.5.0
pydantic-v1  # Compatibility shim for fastapi plugins
httpx==0.25.2
structlog==23.2.0
python-jose[cryptography]==3.3.0

# GPU Monitoring and System Metrics (Phase 5)
pynvml==11.5.0
psutil==5.9.6

# Production Monitoring & Metrics (Phase 5)
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Additional Celery Dependencies for Production (Phase 5)
billiard==4.2.0
kombu==5.3.4

# JSON Processing and Data Validation
orjson==3.9.10
pydantic[email]==2.5.0

# Async HTTP Client for webhooks and callbacks
aiohttp==3.9.1

# Background task monitoring
flower==2.0.1